Register Coordinator works (Clean Code)
