--- Kack2App/src/app/models/teacher.ts
@@ -1,5 +1,0 @@
-import { User } from "./user";
-
-export class Teacher extends User{
-
-}
