+++ Kack2App/src/app/models/user.ts
@@ -1,7 +1,7 @@
 import { EnumRoles } from '../services/enums/enums.service';
 import { Facility } from './facility';
 
-export interface IRoles{
+export interface IRoles {
     trainee: boolean;
     coordinator: boolean;
     admin: boolean;
@@ -14,54 +14,50 @@ export interface IUser {
     displayName: string;
     emailVerified: boolean;
     role: string,
-    name: string; 
+    name: string;
     firstName: string;
- }
+}
 
- export class User implements IUser{
-     _ID!: string;
-     _email!: string;
-     _displayName!: string;
-     _emailVerified!: boolean;
-     _name!: string;
-     _firstName!: string;
-      _role!: string;
-     private _homeFacility = new Facility(); // sollte in Trainee und Teacher ausgelagert werden
+export class User implements IUser {
+    _ID!: string;
+    _email!: string;
+    _displayName!: string;
+    _emailVerified!: boolean;
+    _name!: string;
+    _firstName!: string;
+    _role!: string;
+    private _homeFacility = new Facility(); // sollte in Trainee und Teacher ausgelagert werden
 
-     public get homeFacility() {
-         return this._homeFacility;
-     }
-     public set homeFacility(value) {
-         this._homeFacility = value;
-     }
-     public get role(): string {
+    public get homeFacility() {
+        return this._homeFacility;
+    }
+    public set homeFacility(value) {
+        this._homeFacility = value;
+    }
+    public get role(): string {
         return this._role;
     }
     public set role(value: string) {
         this._role = value;
     }
-
     public get ID(): string {
         return this._ID;
     }
     public set ID(value: string) {
         this._ID = value;
     }
-
     public get email(): string {
         return this._email;
     }
     public set email(value: string) {
         this._email = value;
     }
-
     public get displayName(): string {
         return this._displayName;
     }
     public set displayName(value: string) {
         this._displayName = value;
     }
-
     public get emailVerified(): boolean {
         return this._emailVerified;
     }
@@ -74,7 +70,6 @@ export interface IUser {
     public set name(value: string) {
         this._name = value;
     }
-
     public get firstName(): string {
         return this._firstName;
     }
@@ -82,4 +77,20 @@ export interface IUser {
         this._firstName = value;
     }
 
+}
+
+export class Teacher extends User{
+    _role: string = EnumRoles.teacher;
+}
+
+export class Trainee extends User{
+    _role = EnumRoles.trainee; 
+    age!: number; 
+    DATE_OF_BIRTH!: Date;
+    START_OF_TRAINING!: Date;
+    endOfTraining!: Date
+}
+
+export class Coordinator extends User{
+    _role: string = EnumRoles.coordinator;
 }
